pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idea-manager-bot"
version = "0.1.0"
description = "Telegram bot for capturing and developing ideas inside MYCODEX workspace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "openai>=1.0.0",
  "python-dotenv>=1.0.1",
  "certifi>=2024.0.0",
]

[project.scripts]
idea-manager-bot = "idea_manager_bot.bot:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
